refactor : 네이밍 수정(useModalStatus > useModals)
